backend/requirements.txt
fastapi
uvicorn
opencv-python-headless
pydantic
torch
torchvision
pymongo
python-dotenv
pandas
requests
python-multipart
tensorflow
matplotlib
ultralytics